+++ libnitrate-emit/include/nitrate-emit/Config.h
