+++ libnitrate-emit/src/core/DefaultConfig.cc
@@ -29,138 +29,15 @@
 ///                                                                          ///
 ////////////////////////////////////////////////////////////////////////////////
 
-#include <llvm-14/llvm/MC/TargetRegistry.h>
-#include <llvm-14/llvm/Support/ManagedStatic.h>
-#include <llvm-14/llvm/Support/TargetSelect.h>
-#include <nitrate-core/Lib.h>
-#include <quix-codegen/Lib.h>
-#include <sys/resource.h>
+#define __QUIX_CODEGEN_IMPL__
 
-#include <atomic>
-#include <iostream>
+#include <nitrate-emit/Config.h>
 
-#include "core/LibMacro.h"
+#include <vector>
 
-static std::atomic<size_t> qcode_lib_ref_count = 0;
-
-static void increase_stack_size() {
-  const rlim_t kStackSize = 64 * 1024 * 1024;  // min stack size = 64 MB
-  struct rlimit rl;
-  int result;
-
-  result = getrlimit(RLIMIT_STACK, &rl);
-  if (result == 0) {
-    if (rl.rlim_cur < kStackSize) {
-      rl.rlim_cur = kStackSize;
-      result = setrlimit(RLIMIT_STACK, &rl);
-      if (result != 0) {
-        qcore_panicf("setrlimit returned result = %d\n", result);
-      }
-    }
-  }
+namespace codegen::conf {
+  std::vector<qcode_setting_t> default_settings = {
+      {QCK_CRASHGUARD, QCV_ON},
+      {QCV_FASTERROR, QCV_OFF},
+  };
 }
-
-static bool InitializeLLVM() {
-#ifdef LLVM_SUUPORT_ALL_TARGETS
-  llvm::InitializeAllTargetInfos();
-  llvm::InitializeAllTargets();
-  llvm::InitializeAllTargetMCs();
-  llvm::InitializeAllAsmParsers();
-  llvm::InitializeAllAsmPrinters();
-
-  /* Check if LLVM is initialized */
-  if (llvm::TargetRegistry::targets().empty()) {
-    std::cerr << "error: LLVM initialization failed" << std::endl;
-    return false;
-  }
-#else
-#warning "Building LIBQUIXCC without support for ANY LLVM targets!!"
-#endif
-
-  return true;
-}
-
-static void DoinitializeLLVM() { llvm::llvm_shutdown(); }
-
-static bool do_init() {
-  increase_stack_size();
-  if (!InitializeLLVM()) {
-    return false;
-  }
-
-  return true;
-}
-
-static void do_deinit() { DoinitializeLLVM(); }
-
-LIB_EXPORT bool qcode_lib_init() {
-  if (qcode_lib_ref_count++ > 1) {
-    return true;
-  }
-
-  if (!qcore_lib_init()) {
-    return false;
-  }
-
-  return do_init();
-}
-
-LIB_EXPORT void qcode_lib_deinit() {
-  if (--qcode_lib_ref_count > 0) {
-    return;
-  }
-
-  do_deinit();
-
-  qcore_lib_deinit();
-}
-
-LIB_EXPORT const char* qcode_lib_version() {
-  static const char* version_string =
-
-      "[" __TARGET_VERSION
-      "] ["
-
-#if defined(__x86_64__) || defined(__amd64__) || defined(__amd64) || defined(_M_X64) || \
-    defined(_M_AMD64)
-      "x86_64-"
-#elif defined(__i386__) || defined(__i386) || defined(_M_IX86)
-      "x86-"
-#elif defined(__aarch64__)
-      "aarch64-"
-#elif defined(__arm__)
-      "arm-"
-#else
-      "unknown-"
-#endif
-
-#if defined(__linux__)
-      "linux-"
-#elif defined(__APPLE__)
-      "macos-"
-#elif defined(_WIN32)
-      "win32-"
-#else
-      "unknown-"
-#endif
-
-#if defined(__clang__)
-      "clang] "
-#elif defined(__GNUC__)
-      "gnu] "
-#else
-      "unknown] "
-#endif
-
-#if NDEBUG
-      "[release]"
-#else
-      "[debug]"
-#endif
-
-      ;
-
-  return version_string;
-}
-
-LIB_EXPORT const char* qcode_strerror() { return ""; }
