+++ libnitrate-emit/src/core/Lib.cc
@@ -29,59 +29,138 @@
 ///                                                                          ///
 ////////////////////////////////////////////////////////////////////////////////
 
-#ifndef __QUIX_CODEGEN_LIB_H__
-#define __QUIX_CODEGEN_LIB_H__
+#include <llvm-14/llvm/MC/TargetRegistry.h>
+#include <llvm-14/llvm/Support/ManagedStatic.h>
+#include <llvm-14/llvm/Support/TargetSelect.h>
+#include <nitrate-core/Lib.h>
+#include <nitrate-emit/Lib.h>
+#include <sys/resource.h>
 
-#include <quix-codegen/Config.h>
-#include <stdbool.h>
+#include <atomic>
+#include <iostream>
 
-#ifdef __cplusplus
-extern "C" {
-#endif
+#include "core/LibMacro.h"
 
-/**
- * @brief Initialize the library.
- *
- * @return true if the library was initialized successfully.
- * @note This function is thread-safe.
- * @note The library is reference counted, so it is safe to call this function
- * multiple times. Each time will not reinitialize the library, but will
- * increment the reference count.
- */
-bool qcode_lib_init();
+static std::atomic<size_t> qcode_lib_ref_count = 0;
 
-/**
- * @brief Deinitialize the library.
- *
- * @note This function is thread-safe.
- * @note The library is reference counted, so it is safe to call this function
- * multiple times. Each time will not deinitialize the library, but when
- * the reference count reaches zero, the library will be deinitialized.
- */
-void qcode_lib_deinit();
+static void increase_stack_size() {
+  const rlim_t kStackSize = 64 * 1024 * 1024;  // min stack size = 64 MB
+  struct rlimit rl;
+  int result;
 
-/**
- * @brief Get the version of the library.
- *
- * @return The version string of the library.
- * @warning Don't free the returned string.
- * @note This function is thread-safe.
- * @note This function is safe to call before initialization and after deinitialization.
- */
-const char* qcode_lib_version();
-
-/**
- * @brief Get the last error message from the current thread.
- *
- * @return The last error message from the current thread.
- * @warning Don't free the returned string.
- * @note This function is thread-safe.
- * @note This function is safe to call before initialization and after deinitialization.
- */
-const char* qcode_strerror();
-
-#ifdef __cplusplus
+  result = getrlimit(RLIMIT_STACK, &rl);
+  if (result == 0) {
+    if (rl.rlim_cur < kStackSize) {
+      rl.rlim_cur = kStackSize;
+      result = setrlimit(RLIMIT_STACK, &rl);
+      if (result != 0) {
+        qcore_panicf("setrlimit returned result = %d\n", result);
+      }
+    }
+  }
 }
+
+static bool InitializeLLVM() {
+#ifdef LLVM_SUUPORT_ALL_TARGETS
+  llvm::InitializeAllTargetInfos();
+  llvm::InitializeAllTargets();
+  llvm::InitializeAllTargetMCs();
+  llvm::InitializeAllAsmParsers();
+  llvm::InitializeAllAsmPrinters();
+
+  /* Check if LLVM is initialized */
+  if (llvm::TargetRegistry::targets().empty()) {
+    std::cerr << "error: LLVM initialization failed" << std::endl;
+    return false;
+  }
+#else
+#warning "Building LIBQUIXCC without support for ANY LLVM targets!!"
 #endif
 
-#endif  // __QUIX_CODEGEN_LIB_H__
+  return true;
+}
+
+static void DoinitializeLLVM() { llvm::llvm_shutdown(); }
+
+static bool do_init() {
+  increase_stack_size();
+  if (!InitializeLLVM()) {
+    return false;
+  }
+
+  return true;
+}
+
+static void do_deinit() { DoinitializeLLVM(); }
+
+LIB_EXPORT bool qcode_lib_init() {
+  if (qcode_lib_ref_count++ > 1) {
+    return true;
+  }
+
+  if (!qcore_lib_init()) {
+    return false;
+  }
+
+  return do_init();
+}
+
+LIB_EXPORT void qcode_lib_deinit() {
+  if (--qcode_lib_ref_count > 0) {
+    return;
+  }
+
+  do_deinit();
+
+  qcore_lib_deinit();
+}
+
+LIB_EXPORT const char* qcode_lib_version() {
+  static const char* version_string =
+
+      "[" __TARGET_VERSION
+      "] ["
+
+#if defined(__x86_64__) || defined(__amd64__) || defined(__amd64) || defined(_M_X64) || \
+    defined(_M_AMD64)
+      "x86_64-"
+#elif defined(__i386__) || defined(__i386) || defined(_M_IX86)
+      "x86-"
+#elif defined(__aarch64__)
+      "aarch64-"
+#elif defined(__arm__)
+      "arm-"
+#else
+      "unknown-"
+#endif
+
+#if defined(__linux__)
+      "linux-"
+#elif defined(__APPLE__)
+      "macos-"
+#elif defined(_WIN32)
+      "win32-"
+#else
+      "unknown-"
+#endif
+
+#if defined(__clang__)
+      "clang] "
+#elif defined(__GNUC__)
+      "gnu] "
+#else
+      "unknown] "
+#endif
+
+#if NDEBUG
+      "[release]"
+#else
+      "[debug]"
+#endif
+
+      ;
+
+  return version_string;
+}
+
+LIB_EXPORT const char* qcode_strerror() { return ""; }
