+++ libnitrate-emit/src/core/LibMacro.h
