+++ libnitrate-emit/src/transcode/C11.cc
