+++ libnitrate-emit/src/transcode/CSharp.cc
