+++ libnitrate-emit/src/transcode/CXX11.cc
