+++ libnitrate-emit/src/transcode/Rust.cc
