+++ libnitrate-emit/src/transcode/TS.cc
