+++ libquix/CMakeLists.txt
@@ -34,15 +34,15 @@ target_include_directories(quix PUBLIC src "include"
   ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-seq/include
   ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-parser/include
   ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-ir/include
-  ${CMAKE_CURRENT_SOURCE_DIR}/../libquix-codegen/include)
+  ${CMAKE_CURRENT_SOURCE_DIR}/../libnitrate-emit/include)
 target_compile_options(quix PRIVATE ${PROJECT_DEBUG_FLAGS})
 target_link_libraries(quix PUBLIC nitrate-core
   nitrate-lexer
   nitrate-seq
   nitrate-parser
   nitrate-ir
-  quix-codegen)
-add_dependencies(quix nitrate-core nitrate-lexer nitrate-seq nitrate-parser nitrate-ir quix-codegen)
+  nitrate-emit)
+add_dependencies(quix nitrate-core nitrate-lexer nitrate-seq nitrate-parser nitrate-ir nitrate-emit)
 set_target_properties(quix PROPERTIES OUTPUT_NAME quix)
 
 if(BUILD_TESTING)
