+++ qpkg/CMakeLists.txt
@@ -29,7 +29,7 @@ target_include_directories(qpkg PUBLIC src ${CMAKE_SOURCE_DIR}/libnitrate-core/i
   ${CMAKE_SOURCE_DIR}/libnitrate-seq/include
   ${CMAKE_SOURCE_DIR}/libnitrate-parser/include
   ${CMAKE_SOURCE_DIR}/libnitrate-ir/include
-  ${CMAKE_SOURCE_DIR}/libquix-codegen/include
+  ${CMAKE_SOURCE_DIR}/libnitrate-emit/include
   ${CMAKE_CURRENT_SOURCE_DIR}/third_party/yaml-cpp-0.8.0/include
   ${CMAKE_CURRENT_SOURCE_DIR}/third_party/quixd/include)
 target_compile_options(qpkg PRIVATE
@@ -37,7 +37,7 @@ target_compile_options(qpkg PRIVATE
   -D__TARGET_VERSION=${__TARGET_VERSION}
 )
 target_link_libraries(qpkg PUBLIC libboost_filesystem.a
-  quix-codegen
+  nitrate-emit
   nitrate-ir
   nitrate-parser
   nitrate-lexer
@@ -45,7 +45,7 @@ target_link_libraries(qpkg PUBLIC libboost_filesystem.a
   nitrate-core
   yaml-cpp
   quixd_main)
-add_dependencies(qpkg quix-codegen
+add_dependencies(qpkg nitrate-emit
   nitrate-ir
   nitrate-parser
   nitrate-lexer
